Fix: #506 - Moved logic of VaadinSessionScopeActivationPolicy to VaadinSessionScopedContext
